requirements.txt
Pillow==9.0.1
simplemma==0.5.0
pydantic==1.9.0
fastapi==0.74.1
uvicorn[standard]==0.17.5
pytesseract==0.3.9
python-multipart==0.0.5
opencv-python==4.5.5.62
pytest>=7.1.0
numpy>=1.22.2
